// src/app/shared/z_data-entry-item-input/data-entry-item-type.enum.ts
export enum DataEntryItemType{
    TimelogEntry = "TIMELOG_ENTRY",
    FinancialEntry = "FINANCIAL_ENTRY",
    ActionItemEntry = "ACTION_ITEM_ENTRY",
    ReminderEntry = "REMINDER_ENTRY",
    ScheduledEventEntry = "SCHEDULED_EVENT_ENTRY",
    DietaryEntry = "DIETARY_ENTRY",
    FeelingEntry = "FEELING_ENTRY",
    WeightlogEntry = "WEIGHTLOG_ENTRY",
    CountEntry = "COUNT_ENTRY",
    RuleConditionEntry = "RULE_CONDITION_ENTRY",
    HealthSymptomEntry = "HEALTH_SYMPTOM_ENTRY",
    // ExerciseEntry = ""
    CustomDefintion = "CUSTOM_DEFINITION",
}